requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.1
async-timeout==4.0.3
asyncpg==0.28.0
asynctest==0.13.0
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
frozenlist==1.3.3
greenlet==3.0.1
h11==0.14.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
magic-filter==1.0.11
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
outcome==1.2.0
psycopg2-binary==2.9.9
PySocks==1.7.1
pytz==2023.3
PyVirtualDisplay==3.0
python-dotenv==1.0.0
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.22
SQLAlchemy-ImageAttach==1.1.0
trio==0.22.2
trio-websocket==0.10.3
typing-extensions==4.7.1
urllib3==2.0.4
Wand==0.6.11
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
